attributes dataset

+++ hpc_test/hpc_test.py
@@ -17,7 +17,7 @@ def get_args():
     parser.add_argument('--workers', type=int, default=2, help='number of workers in data loader')
     parser.add_argument('--print_every', type=int, default=500, help='print losses every N iteration')
 
-    parser.add_argument('--lr', type=int, default=0.001, help='learning rate')
+    parser.add_argument('--lr', type=float, default=0.001, help='learning rate')
     parser.add_argument('--opt', type=str, default='SGD', choices=['SGD', 'Adam'], help = 'optimizer used for training')
     parser.add_argument('--use_norm', action='store_true', help='use normalization layers in model')
     parser.add_argument('--feat', type=int, default=16, help='number of features in model')
